Smart Alarm v2.0 Initial Commit
Restructured the program so that the logic is center piece. The GUI
holds all the tkinter commands, and Alarm API holds just the methods for
Google and Weather. The alarm has been rewritten so that the subprocess
is monitored in the GUITimer method rather than in a separate
multiprocessor.Process.